+++ packages/use-request/src/utils/index.ts
@@ -1,13 +1,1 @@
-export function isDocumentVisible(): boolean {
-  if (typeof document !== 'undefined' && typeof document.visibilityState !== 'undefined') {
-    return document.visibilityState !== 'hidden';
-  }
-  return true;
-}
-
-export function isOnline(): boolean {
-  if (typeof navigator.onLine !== 'undefined') {
-    return navigator.onLine;
-  }
-  return true;
-}
+export * from './utils';
