+++ yarn-error.log
@@ -1,0 +1,96 @@
+Arguments: 
+  /usr/local/bin/node /usr/local/bin/yarn
+
+PATH: 
+  /usr/local/bin:/usr/local/sbin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Library/Apple/usr/bin
+
+Yarn version: 
+  1.22.10
+
+Node version: 
+  14.17.1
+
+Platform: 
+  darwin x64
+
+Trace: 
+  Error: read ECONNRESET
+      at TLSWrap.onStreamRead (internal/stream_base_commons.js:209:20)
+
+npm manifest: 
+  {
+    "name": "ahooks",
+    "private": true,
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/alibaba/hooks.git"
+    },
+    "scripts": {
+      "init": "rm -rf node_modules && npm install && npm run clean && npm run bootstrap && npm run build",
+      "start": "npm run dev",
+      "dev": "dumi dev",
+      "bootstrap": "lerna bootstrap",
+      "clean": "lerna clean --yes",
+      "build": "lerna run build",
+      "test": "jest",
+      "lint": "eslint --ignore-pattern **/__tests__/* --ignore-pattern **/demo/* \"packages/*/src/**/*.{ts,tsx}\"",
+      "build:doc": "dumi build",
+      "pub:doc": "surge ./dist --domain ahooks.js.org",
+      "pub:doc-gitee": "cd ./dist && rm -rf .git && touch .spa && touch .nojekyll && git init && git remote add origin git@gitee.com:ahooks/ahooks.git && git add -A && git commit -m \"publish docs\" && git push origin master -f",
+      "pub": "lerna publish"
+    },
+    "husky": {
+      "hooks": {
+        "pre-commit": "pretty-quick --staged"
+      }
+    },
+    "devDependencies": {
+      "@babel/cli": "^7.10.1",
+      "@babel/core": "^7.10.2",
+      "@testing-library/react": "^10.0.4",
+      "@testing-library/react-hooks": "^2.0.1",
+      "@types/jest": "^24.9.1",
+      "@types/lodash.debounce": "^4.0.6",
+      "@types/lodash.isequal": "^4.5.5",
+      "@types/lodash.throttle": "^4.1.6",
+      "@umijs/fabric": "^2.1.0",
+      "@umijs/plugin-sass": "^1.1.1",
+      "antd": "^4.3.3",
+      "babel-loader": "^8.1.0",
+      "babel-plugin-import": "^1.12.0",
+      "babel-plugin-transform-async-to-promises": "^0.8.15",
+      "del": "^5.1.0",
+      "dumi": "^1.1.7",
+      "enzyme": "^3.10.0",
+      "eslint": "^7.2.0",
+      "eslint-plugin-react-hooks": "^4.0.8",
+      "gulp": "^4.0.2",
+      "gulp-babel": "^8.0.0",
+      "gulp-typescript": "^6.0.0-alpha.1",
+      "husky": "^4.2.5",
+      "jest": "^26.0.1",
+      "jest-fetch-mock": "^3.0.3",
+      "lerna": "^2.11.0",
+      "mockdate": "^3.0.2",
+      "mockjs": "^1.1.0",
+      "prettier": "^2.0.5",
+      "pretty-quick": "^2.0.1",
+      "react": "^16.8.6",
+      "react-dom": "^16.8.6",
+      "react-drag-listview": "^0.1.6",
+      "react-test-renderer": "^16.13.1",
+      "surge": "^0.21.3",
+      "ts-jest": "^26.1.0",
+      "typescript": "^3.3.3",
+      "umi-request": "^1.2.18",
+      "webpack": "^4.43.0",
+      "webpack-cli": "^3.3.10",
+      "webpack-merge": "^4.2.2"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  No lockfile
